requirements.txt
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
docopt==0.6.2
h11==0.12.0
idna==3.3
outcome==1.1.0
Pillow==8.4.0
pycparser==2.21
pyOpenSSL==21.0.0
PyYAML==6.0
requests==2.26.0
schedule==1.1.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wsproto==1.0.0
